+++ src/Umbraco.Tests/Umbraco.Tests.csproj
@@ -231,7 +231,6 @@
     <Compile Include="Services\CachedDataTypeServiceTests.cs" />
     <Compile Include="TestHelpers\Entities\MockedPropertyTypes.cs" />
     <Compile Include="TestHelpers\Entities\MockedUserGroup.cs" />
-    <Compile Include="TryConvertToTests.cs" />
     <Compile Include="UdiTests.cs" />
     <Compile Include="UmbracoExamine\RandomIdRAMDirectory.cs" />
     <Compile Include="UmbracoExamine\UmbracoContentIndexerTests.cs" />
